+++ e2e/vm/additional_disk_test.go
@@ -42,8 +42,9 @@ var testAdditionalDisk = func(o *option.Option) {
 			imageOutput := command.StdoutAsLines(o, "images", "--format", "{{.Name}}")
 			gomega.Expect(imageOutput).Should(gomega.ContainElement(savedImage))
 
+			// Changed in nerdctl v1.5 to [<name>] https://github.com/containerd/nerdctl/commit/11d80f274257c064924f40bd007756110d863a16
 			psOutput := command.StdoutAsLines(o, "ps", "--all", "--format", "{{.Names}}")
-			gomega.Expect(psOutput).Should(gomega.ContainElement(containerName))
+			gomega.Expect(psOutput).Should(gomega.ContainElement("[" + containerName + "]"))
 
 			volumeOutput := command.StdoutAsLines(o, "volume", "ls", "--format", "{{.Name}}")
 			gomega.Expect(volumeOutput).Should(gomega.ContainElement(volumeName))